+++ .env
@@ -7,7 +7,7 @@ DB_DATABASE="crypto_nft"
 
 WS_URL="wss://speedy-nodes-nyc.moralis.io/c3e30c53ff79f6a1980cd9ea/eth/rinkeby/ws"
 
-TOKENMANAGER_CONTRACT_ADDRESS="0x3dd2757C0344602E982B646463019c7e5fa7AF3A"
+TOKENMANAGER_CONTRACT_ADDRESS="0x5F6C14a5AC3648Fbc81B2221CF4560b23c354BC1"
 
 DEVELOPMENT_URL="http://localhost:8080/"
 PRODUCTION_URL="https://api.dialogbox.com/"
